Simple demo of thread local issue
When it works curl localhost:8080 returns a stream of HTML (one
message every 5 seconds), and all the logs show that the
request attributes are available to the message converter.

When you uncomment the (correct) implementation of the reset
method in the thread local accessor it still works, but the
request attributes are null because they are reset by reactor
too soon.